requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.27.0

# Google APIs
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.108.0

# Cloud Infrastructure
modal>=0.64.0
uvicorn>=0.29.0
fastapi[standard]>=0.111.0

# Utilities
python-dateutil>=2.8.2
esprima>=4.0.1  # JS syntax checker used by execution/diagnose.py

# Attorney Mapper
serpapi>=2.4.2
geopy>=2.4.1
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.25.0
